requirements/test.txt
# pyup: ignore file
pytest
pytest-travis-fold
pytest-cov
pytest-django
future

git+https://github.com/moreati/python-zxcvbn
